requirements.txt
fastapi[all]
alembic
sqlalchemy
psycopg2
python-dotenv
pydub
